Added a dash cooldown
Basically, whenever the player dashes, the cooldown starts. If the player dashes during the cooldown, the cooldown starts over. The player can have a certain amount of dashes.

For example: If the player has two dashes, they could dash once and the cooldown for that dash would begin. They are able to dash during this cooldown, since they have a second dash.
However, if they use that second dash before the cooldown for the first one is over, then the cooldown starts over and they won't be able to dash, since they've used their last dash. Once the restarted cooldown ends, they will have one dash, and must wait for the next dash.

This is how Ultrakill does it. Let me know if we should change it to just completely recharge dashes after one cooldown.