Fix to make the particles from delete themselves. PowerUps now drop from monster kills.
How does it work ?
There is a powerup manager, it will be called when a monster is destroyed by the scrip "OnDestroyCheckForDrops" which is placed on the monster.

The manager checks if a random number is lower than the drop rate. If it is, it will drop a random buff from a list present in the manager.