Damage Numbers and a second punch animation for swarmer
I added an overload to TakeDamage which added one method to IDamageble, this caused me to have this method introduced to all IDamageble scripts and thats why it looks like I touched everything.

Font size, font itself and colors should be adjusted in the future to make it look better but functionality is there.